refactor: Extract search with lunr
